add support for screenlayout files
On a per-hostname basis, a screenlayout file can be be executed when
going from docked to undocked or vice-versa. Some various timing-related
intervals have changed also. Some work may need to be done in the future
to deal with cases where one host could be docked to one of several
docking stations. Perhaps multiple screenlayout files could be searched
and executed one at a time until one exits nonzero, which should work
since xrandr will not allow the monitor configuration to be changed to
an invalid one.