disavow: add --
The -- causes the argument parser to stop parsing arguments. This is
useful when a substring of the command to run is also a valid argument
to disavow.

For example 'disavow foo -p' would result in the command 'foo' running,
and it's PID being printed thereafter. With the new behavior, 'disavow
-- foo -p' would result in 'foo -p' being run, and its PID would not be
printed, and 'disavow -p -- foo -p' would cause 'foo -p' to be run, and
it's PID to be printed after disowning it.