Added require-sudo
If a task requires sudo access, the user can either provide it on argv
to taskutil, in a file, or can be prompted interactively for it. On
tasks with require-sudo marked as true, the environment variable
TASKUTIL_SUDO_PASSWORD will contain the sudo password, which can then be
used with sudo -S.

This commit also provides improvements to the way do_run_script works,
namely in that scripts can no longer write directly to standard error
nor can they take control of standard input.